Pre-review changes and update to use double tick parameter for EmStd.

+++ table-api/src/main/java/io/deephaven/api/updateby/spec/EmStdSpec.java
@@ -42,11 +42,11 @@ public abstract class EmStdSpec extends UpdateBySpecBase {
         return of(WindowScale.ofTime(timestampCol, emaDuration));
     }
 
-    public static EmStdSpec ofTicks(OperationControl control, long tickWindow) {
+    public static EmStdSpec ofTicks(OperationControl control, double tickWindow) {
         return of(control, WindowScale.ofTicks(tickWindow));
     }
 
-    public static EmStdSpec ofTicks(long tickWindow) {
+    public static EmStdSpec ofTicks(double tickWindow) {
         return of(WindowScale.ofTicks(tickWindow));
     }
